+++ src/main/java/de/codescape/jira/plugins/multiplesubtasks/MultipleSubtasksConstants.java
@@ -1,0 +1,13 @@
+package de.codescape.jira.plugins.multiplesubtasks;
+
+/**
+ * Globally used constants that are reused among Multiple Subtasks for Jira classes go into this class.
+ */
+public class MultipleSubtasksConstants {
+
+    /**
+     * Plugin key that identifies the Multiple Subtasks for Jira plugin.
+     */
+    public static final String MULTIPLE_SUBTASKS_PLUGIN_KEY = "de.codescape.jira.plugins.multiple-subtasks";
+
+}